tutorials/pyproject.toml
[project]
name = "tutorials"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.1.9",
    "langgraph>=0.6.5",
    "langsmith>=0.4.14",
    "notebook>=7.4.5",
    "python-dotenv>=1.1.1",
]
